Require GitHub personal access token as option
We are now requiring a GitHub personal access token to be explicitly
passed as an option to this library and it no longer attempts to
auto-magically read a token from the environment variable
`GITHUB_PERSONAL_ACCESS_TOKEN`. This allows end users of this
library to decide how and where they store their GitHub personal
access token e.g. environment variable, file, outer space.