Replace nodegit dependency with dugite
We're ditching NodeGit as it is tricky to use and has very poor error
handling. Instead I'm rewriting the git helper to use dugite, which
directly calls the git command line interface. In the process, I'm
dropping the use of classes as they don't provide any benefits in
terms of the structure or usage of this helper - simple modules ftw.