Add internal helper for handling dugite results
This is an internal helper to handle the result from dugite's
`GitProcess.exec` method. If the git command did not execute
successfully it will throw an error. In order to aid the end
user with debugging, the error message includes the full git
command that was executed as well as the raw error message
from the git CLI.