Add Jest mock for dugite module
The tests for the methods in this module will focus on testing that
they correctly handle valid and invalid options and provide appropriate
return values. We don't want to be testing dugite itself so we're
mocking the whole module, then each suite can mock the resolved value
from `GitProcess.exec` as they require.