Add changelog module
The changelog used to be a part of the `PackageJson` class, but now
we've refactored it into its own module in order to separate the
concerns of manipulating the `package.json` document and keeping
track of the changes that have been made. This also allows us to test
the changelog methods independently.