Refactor PackageJson class into a simple module
We have refactored the `PackageJson` class into a simple module that
exposes a single method, `loadPackageJson`. This method loads a
`package.json` file in to memory so that it can be manipulated.

After reading the specified `package.json` file, the `loadPackageJson`
method returns a collection of methods that can be used for changing
the `package.json` document and writing those changes back to disk.

Methods returned by `loadPackageJson`:

- getChangelog
- getDocument
- hasChangesToWrite
- writeChanges
- getField
- setField
- requireDependency
- removeDependency
- requireScript

This module uses the new `changelog` module, as well as the
`deepCloneObject` and `formatObjectAsJson` helpers.

In the course of refactoring this module we have made the behaviour
of the methods more consistent and improved the error handling.